+++ src/main/java/oshi/hardware/platform/linux/LinuxHardwareAbstractionLayer.java
@@ -19,6 +19,11 @@ package oshi.hardware.platform.linux;
 import java.io.IOException;
 import java.util.List;
 
+import javax.json.Json;
+import javax.json.JsonArrayBuilder;
+import javax.json.JsonBuilderFactory;
+import javax.json.JsonObject;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -30,7 +35,6 @@ import oshi.hardware.PowerSource;
 import oshi.software.os.OSFileStore;
 import oshi.software.os.linux.LinuxFileSystem;
 import oshi.util.FileUtil;
-import oshi.util.ParseUtil;
 
 /**
  * @author alessandro[at]perucchi[dot]org
@@ -44,6 +48,8 @@ public class LinuxHardwareAbstractionLayer implements HardwareAbstractionLayer {
     private CentralProcessor processor;
 
     private GlobalMemory _memory;
+
+    private JsonBuilderFactory jsonFactory = Json.createBuilderFactory(null);
 
     @Override
     public GlobalMemory getMemory() {
@@ -124,14 +130,21 @@ public class LinuxHardwareAbstractionLayer implements HardwareAbstractionLayer {
     }
 
     @Override
-    public String toJSON() {
-        StringBuilder sb = new StringBuilder("{");
-        sb.append("\"processor\":").append(getProcessor().toJSON()).append(",");
-        sb.append("\"memory\":").append(getMemory().toJSON()).append(",");
-        sb.append("\"powerSources\":").append(ParseUtil.toJsonArray(getPowerSources())).append(",");
-        sb.append("\"fileStores\":").append(ParseUtil.toJsonArray(getFileStores())).append(",");
-        sb.append("\"displays\":").append(ParseUtil.toJsonArray(getDisplays()));
-        return sb.append("}").toString();
+    public JsonObject toJSON() {
+        JsonArrayBuilder powerSourceArrayBuilder = jsonFactory.createArrayBuilder();
+        for (PowerSource powerSource : getPowerSources()) {
+            powerSourceArrayBuilder.add(powerSource.toJSON());
+        }
+        JsonArrayBuilder fileStoreArrayBuilder = jsonFactory.createArrayBuilder();
+        for (OSFileStore fileStore : getFileStores()) {
+            fileStoreArrayBuilder.add(fileStore.toJSON());
+        }
+        JsonArrayBuilder displayArrayBuilder = jsonFactory.createArrayBuilder();
+        for (Display display : getDisplays()) {
+            displayArrayBuilder.add(display.toJSON());
+        }
+        return jsonFactory.createObjectBuilder().add("processor", getProcessor().toJSON())
+                .add("memory", getMemory().toJSON()).add("powerSources", powerSourceArrayBuilder.build())
+                .add("fileStores", fileStoreArrayBuilder.build()).add("displays", displayArrayBuilder.build()).build();
     }
-
 }