+++ src/main/java/oshi/hardware/platform/mac/MacPowerSource.java
@@ -19,6 +19,10 @@ package oshi.hardware.platform.mac;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.json.Json;
+import javax.json.JsonBuilderFactory;
+import javax.json.JsonObject;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,7 +36,6 @@ import oshi.jna.platform.mac.CoreFoundation.CFArrayRef;
 import oshi.jna.platform.mac.CoreFoundation.CFDictionaryRef;
 import oshi.jna.platform.mac.CoreFoundation.CFTypeRef;
 import oshi.jna.platform.mac.IOKit;
-import oshi.util.ParseUtil;
 
 /**
  * A Power Source
@@ -47,6 +50,8 @@ public class MacPowerSource implements PowerSource {
     private double remainingCapacity;
 
     private double timeRemaining;
+
+    private JsonBuilderFactory jsonFactory = Json.createBuilderFactory(null);
 
     public MacPowerSource(String newName, double newRemainingCapacity, double newTimeRemaining) {
         this.name = newName;
@@ -71,12 +76,9 @@ public class MacPowerSource implements PowerSource {
     }
 
     @Override
-    public String toJSON() {
-        StringBuilder sb = new StringBuilder("{");
-        sb.append("\"name\":").append(ParseUtil.jsonQuote(getName())).append(",");
-        sb.append("\"remainingCapacity\":").append(getRemainingCapacity()).append(",");
-        sb.append("\"timeRemaining\":").append(getTimeRemaining());
-        return sb.append("}").toString();
+    public JsonObject toJSON() {
+        return jsonFactory.createObjectBuilder().add("name", getName()).add("remainingCapacity", getRemainingCapacity())
+                .add("timeRemaining", getTimeRemaining()).build();
     }
 
     /**